requirements.txt
openai==1.6.1
python-dotenv==1.0.0
Pillow==10.1.0
pdf2image==1.16.3
